inline traits, add test (WP-874)
