+++ inc/Smartling/Helpers/LoggerSafeTrait.php
@@ -2,6 +2,7 @@
 
 namespace Smartling\Helpers;
 
+use Smartling\Models\LoggerWithStringContext;
 use Smartling\MonologWrapper\MonologWrapper;
 use Smartling\Vendor\Psr\Log\LoggerInterface;
 use Smartling\Vendor\Psr\Log\NullLogger;
@@ -9,7 +10,7 @@ use Smartling\Vendor\Psr\Log\NullLogger;
 trait LoggerSafeTrait {
     private ?LoggerInterface $logger = null;
 
-    public function getLogger(): LoggerInterface {
+    public function getLogger(): LoggerWithStringContext {
         if ($this->logger === null) {
             $this->logger = MonologWrapper::getLogger(static::class);
         }
